Fix race condition when queueing callbacks.
Per https://github.com/AFNetworking/AFNetworking/issues/429, it was possible for the operation's completion block to schedule the success or failure block, but before it executed the operation was deallocated and its weak references were nil'd.

Creating a strong reference within the block ensures that the operation will not yet be deallocated, but also avoids Xcode's warning about strongly capturing self causing a retain cycle.